requirements.txt
discord-py-slash-command~=2.0.4
discord-py~=1.7.3
python-dotenv~=0.18.0
pymongo~=3.11.4